158: create a separate component for inputRow
